feat: support Google Drive link in audio editor
With this commit, using a Google Drive link in the audio editor will bypass the error event listener.

Why? Normally, using a Google Drive link triggers the error event listener because the audio can't be loaded. However, when played through osu-web, it works fine. This commit allows the use of the Google Drive link while disabling the preview.