feat: new element `Group` and vertical flag wrapper option
as the title says, this commit adds a new canvas element called "group". it's not related to bbcode and is only used to help organize related canvas items. it won't be rendered in bbcode.

also added a new option for flag wrappers, so they can now be displayed vertically or horizontally.